services/analyticslayer/pyproject.toml
[project]
name = "analyticslayer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.6",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.0",
]
